+++ cosmic/resource_cosmic_loadbalancer_rule.go
@@ -64,6 +64,7 @@ func resourceCosmicLoadBalancerRule() *schema.Resource {
 				Type:     schema.TypeString,
 				Optional: true,
 				Computed: true,
+				ForceNew: true,
 				ValidateFunc: func(val interface{}, key string) (warns []string, errs []error) {
 					v := val.(string)
 					switch v {