+++ cosmic/resource_cosmic_loadbalancer_rule_test.go
@@ -208,6 +208,83 @@ func TestAccCosmicLoadBalancerRule_updatePorts(t *testing.T) {
 	})
 }
 
+func TestAccCosmicLoadBalancerRule_updateProtocol(t *testing.T) {
+	if COSMIC_SERVICE_OFFERING_1 == "" {
+		t.Skip("This test requires an existing service offering (set it by exporting COSMIC_SERVICE_OFFERING_1)")
+	}
+
+	if COSMIC_TEMPLATE == "" {
+		t.Skip("This test requires an existing instance template (set it by exporting COSMIC_TEMPLATE)")
+	}
+
+	if COSMIC_VPC_NETWORK_OFFERING == "" {
+		t.Skip("This test requires an existing VPC network offering (set it by exporting COSMIC_VPC_NETWORK_OFFERING)")
+	}
+
+	if COSMIC_VPC_OFFERING == "" {
+		t.Skip("This test requires an existing VPC offering (set it by exporting COSMIC_VPC_OFFERING)")
+	}
+
+	var id string
+	var rule cosmic.LoadBalancerRule
+
+	createAttributes := &testAccCheckCosmicLoadBalancerRuleExpectedAttributes{
+		Name:        "terraform-lb",
+		Algorithm:   "roundrobin",
+		PrivatePort: "8080",
+		PublicPort:  "80",
+	}
+
+	updateAttributes := &testAccCheckCosmicLoadBalancerRuleExpectedAttributes{
+		Name:        "terraform-lb",
+		Algorithm:   "roundrobin",
+		Protocol:    "tcp-proxy",
+		PrivatePort: "8080",
+		PublicPort:  "80",
+	}
+
+	resource.ParallelTest(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckCosmicLoadBalancerRuleDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccCosmicLoadBalancerRule_basic(createAttributes),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					testAccCheckCosmicLoadBalancerRuleExist("cosmic_loadbalancer_rule.foo", &id, &rule, true),
+					testAccCheckCosmicLoadBalancerRuleAttributes(&rule, createAttributes),
+					resource.TestCheckResourceAttr(
+						"cosmic_loadbalancer_rule.foo", "name", createAttributes.Name),
+					resource.TestCheckResourceAttr(
+						"cosmic_loadbalancer_rule.foo", "algorithm", createAttributes.Algorithm),
+					resource.TestCheckResourceAttr(
+						"cosmic_loadbalancer_rule.foo", "private_port", createAttributes.PrivatePort),
+					resource.TestCheckResourceAttr(
+						"cosmic_loadbalancer_rule.foo", "public_port", createAttributes.PublicPort),
+				),
+			},
+
+			{
+				Config: testAccCosmicLoadBalancerRule_updateProtocol(updateAttributes),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					testAccCheckCosmicLoadBalancerRuleExist("cosmic_loadbalancer_rule.foo", &id, &rule, true),
+					testAccCheckCosmicLoadBalancerRuleAttributes(&rule, updateAttributes),
+					resource.TestCheckResourceAttr(
+						"cosmic_loadbalancer_rule.foo", "name", updateAttributes.Name),
+					resource.TestCheckResourceAttr(
+						"cosmic_loadbalancer_rule.foo", "algorithm", updateAttributes.Algorithm),
+					resource.TestCheckResourceAttr(
+						"cosmic_loadbalancer_rule.foo", "protocol", updateAttributes.Protocol),
+					resource.TestCheckResourceAttr(
+						"cosmic_loadbalancer_rule.foo", "private_port", updateAttributes.PrivatePort),
+					resource.TestCheckResourceAttr(
+						"cosmic_loadbalancer_rule.foo", "public_port", updateAttributes.PublicPort),
+				),
+			},
+		},
+	})
+}
+
 func testAccCheckCosmicLoadBalancerRuleExist(n string, id *string, rule *cosmic.LoadBalancerRule, shouldChange bool) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 
@@ -254,6 +331,7 @@ type testAccCheckCosmicLoadBalancerRuleExpectedAttributes struct {
 	Algorithm   string
 	Name        string
 	PrivatePort string
+	Protocol    string
 	PublicPort  string
 }
 
@@ -274,6 +352,12 @@ func testAccCheckCosmicLoadBalancerRuleAttributes(rule *cosmic.LoadBalancerRule,
 
 		if rule.Publicport != want.PublicPort {
 			return fmt.Errorf("Bad public port: got %s; want %s", rule.Publicport, want.PublicPort)
+		}
+
+		if want.Protocol != "" {
+			if rule.Protocol != want.Protocol {
+				return fmt.Errorf("Bad protocol: got %s; want %s", rule.Protocol, want.Protocol)
+			}
 		}
 
 		return nil
@@ -438,6 +522,71 @@ resource "cosmic_loadbalancer_rule" "foo" {
 		COSMIC_TEMPLATE,
 		attr.Name,
 		attr.Algorithm,
+		attr.PublicPort,
+		attr.PrivatePort,
+	)
+}
+
+func testAccCosmicLoadBalancerRule_updateProtocol(attr *testAccCheckCosmicLoadBalancerRuleExpectedAttributes) string {
+	return fmt.Sprintf(`
+resource "cosmic_vpc" "foo" {
+  name           = "terraform-vpc"
+  display_text   = "terraform-vpc"
+  cidr           = "10.0.10.0/22"
+  network_domain = "terraform-domain"
+  vpc_offering   = "%s"
+  zone           = "%s"
+}
+
+resource "cosmic_network" "foo" {
+  name             = "terraform-network"
+  cidr             = "10.0.10.0/24"
+  gateway          = "10.0.10.1"
+  network_offering = "%s"
+  vpc_id           = "${cosmic_vpc.foo.id}"
+  zone             = "${cosmic_vpc.foo.zone}"
+}
+
+data "cosmic_network_acl" "default_allow" {
+  filter {
+    name  = "name"
+    value = "default_allow"
+  }
+}
+
+resource "cosmic_ipaddress" "foo" {
+  acl_id = "${data.cosmic_network_acl.default_allow.id}"
+  vpc_id = "${cosmic_vpc.foo.id}"
+}
+
+resource "cosmic_instance" "foo1" {
+  name             = "terraform-server1"
+  display_name     = "terraform-server1"
+  service_offering = "%s"
+  network_id       = "${cosmic_network.foo.id}"
+  template         = "%s"
+  zone             = "${cosmic_vpc.foo.zone}"
+  expunge          = true
+}
+
+resource "cosmic_loadbalancer_rule" "foo" {
+  name          = "%s"
+  ip_address_id = "${cosmic_ipaddress.foo.id}"
+  algorithm     = "%s"
+  protocol      = "%s"
+  network_id    = "${cosmic_network.foo.id}"
+  public_port   = "%s"
+  private_port  = "%s"
+  member_ids    = ["${cosmic_instance.foo1.id}"]
+}`,
+		COSMIC_VPC_OFFERING,
+		COSMIC_ZONE,
+		COSMIC_VPC_NETWORK_OFFERING,
+		COSMIC_SERVICE_OFFERING_1,
+		COSMIC_TEMPLATE,
+		attr.Name,
+		attr.Algorithm,
+		attr.Protocol,
 		attr.PublicPort,
 		attr.PrivatePort,
 	)